+++ src/vgui2/matsys_controls/baseassetpicker.cpp
@@ -715,7 +715,7 @@ bool CAssetCache::AddFilesInDirectory( CachedAssetList_t& list, const char *pSta
 //-----------------------------------------------------------------------------
 bool CAssetCache::ContinueSearchForAssets( AssetList_t hList, float flDuration )
 {
-	CachedAssetList_t& list = m_CachedAssets[ (int)hList ];
+	CachedAssetList_t& list = m_CachedAssets[ (intp)hList ];
 
 	float flStartTime = Plat_FloatTime();
 	while ( list.m_DirectoriesToCheck.Count() )
@@ -754,7 +754,7 @@ bool CAssetCache::ContinueSearchForAssets( AssetList_t hList, float flDuration )
 //-----------------------------------------------------------------------------
 bool CAssetCache::BeginAssetScan( AssetList_t hList, bool bForceRescan )
 {
-	CachedAssetList_t& list = m_CachedAssets[ (int)hList ];
+	CachedAssetList_t& list = m_CachedAssets[ (intp)hList ];
 	if ( bForceRescan )
 	{
 		list.m_bAssetScanComplete = false;
@@ -807,7 +807,7 @@ void CAssetCache::SetUsedAssetList(CUtlVector<AssetUsageInfo_t> &usedAssets)
 AssetList_t CAssetCache::FindAssetList( const char *pAssetType, const char *pSubDir, int nExtCount, const char **ppExt )
 {
 	CachedAssetList_t search( pSubDir, nExtCount, ppExt );
-	int nIndex = m_CachedAssets.Find( search );
+	intp nIndex = m_CachedAssets.Find( search );
 	if ( nIndex == m_CachedAssets.InvalidIndex() )
 	{
 		nIndex = m_CachedAssets.Insert( search );
@@ -826,14 +826,14 @@ CAssetTreeView* CAssetCache::GetFileTree( AssetList_t hList )
 {
 	if ( hList == ASSET_LIST_INVALID )
 		return NULL;
-	return m_CachedAssets[ (int)hList ].m_pFileTree;
+	return m_CachedAssets[ (intp)hList ].m_pFileTree;
 }
 
 int CAssetCache::GetAssetCount( AssetList_t hList ) const
 {
 	if ( hList == ASSET_LIST_INVALID )
 		return 0;
-	return m_CachedAssets[ (int)hList ].m_AssetList.Count();
+	return m_CachedAssets[ (intp)hList ].m_AssetList.Count();
 }
 
 //const CAssetCache::CachedAssetInfo_t& CAssetCache::GetAsset( AssetList_t hList, int nIndex ) const
